==> R/00_all.qmd <==
---
title: "00_all"
author: "Group 16"
format:
  html:
    embed-resources: true
editor: visual
---

# 00_all

### This is our final Quarto Markdown document with all other documents embedded and rendered

## 01_load

```{r}
quarto::quarto_render("../R/01_load.qmd")

file.rename(from = "../R/01_load.html",
             to = "../results/01_load.html" )
```

## 02_clean

```{r}
quarto::quarto_render("../R/02_clean.qmd")

file.rename(from = "../R/02_clean.html",
             to = "../results/02_clean.html" )
```

## 03_augment

```{r}
quarto::quarto_render("../R/03_augment.qmd")

file.rename(from = "../R/03_augment.html",
             to = "../results/03_augment.html" )
```

## 04_describe

```{r}
quarto::quarto_render("../R/04_describe.qmd")

file.rename(from = "../R/04_describe.html",
             to = "../results/04_describe.html" )
```

## 05_analysis_1

```{r}
quarto::quarto_render("../R/05_analysis_1.qmd")

file.rename(from = "../R/05_analysis_1.html",
             to = "../results/05_analysis_1.html" )
```

## 06_analysis_2

```{r}
quarto::quarto_render("../R/06_analysis_2.qmd")

file.rename(from = "../R/06_analysis_2.html",
             to = "../results/06_analysis_2.html" )
```

## 07_analysis_3

```{r}
quarto::quarto_render("../R/07_analysis_3.qmd")

file.rename(from = "../R/07_analysis_3.html",
             to = "../results/07_analysis_3.html" )
```
