==> R/03_augment.qmd <==
---
format:
  html: {}
---

# Augmentation of the data

1.  Load all important library's

```{r}
library(tidyverse)
```

2.  We load the the cleaned and transformed data from the TSV file "02_dat_clean.tsv" into an R data frame named **`data`**.

```{r}
#| message: false
# Loading the clean data
data <- read_tsv("../data/02_dat_clean.tsv")
data_r <- read_tsv("../data/02_raw_clean.tsv")
```

### Diagnosis

Looking at the data, it can be seen that one of the categorical variables contains information on both presence and absence of heart disease, and the type of heart disease which was diagnosed. To do the analysis, a new variable was created, which only indicates presence or absence of heart disease.

```{r}
data <- data |> 
  mutate(diagnose = case_when(
    num == 0 ~ 'healthy',
    num >= 1 ~ 'sick'
  ))
data_r <- data_r |> 
  mutate(diagnose = case_when(
    num == 0 ~ 0,
    num >= 1 ~ 1
  ))

```

The new data is saved to the datafile

```{r}
write_tsv(data, "../data/03_dat_aug.tsv")
write_tsv(data_r, "../data/03_raw_aug.tsv")
```
