==> R/04_describe.qmd <==
---
format:
  html: {}
---

# Description of the data

### Introduction

In this section, we'll take a closer look at various details about our heart disease dataset and provide a detailed explanation of the variables.

### Loading the data

```{r}
#| message: false
library(tidyverse)
```

```{r}
# Loading the clean data
data <- read_tsv("../data/02_dat_clean.tsv")
data
```

### Dataset Overview

After loading the dataset into R, it revealed 14 variables and 303 observations. It provides key information, including details about age, gender, and various aspects of heart health of certain patients. The 'num' variable indicates the presence of heart disease, ranging from 0 (no presence) to 4. During augmentation of the data, a new variable was added, which only indicates if a patient is diagnosed or not, called 'diagnosed'. Also, patient names and social security numbers were replaced with dummy values for privacy. In the following, we will take a deeper look at the variable information to make the resulting interpretation of the dataset easier.

### Variable Information

The upcoming variable documentation consists of two different sub-parts. The documentation will be split into 'categorical variables' and 'continuous variables.' 'Categorical variables' contain a finite number of categories or distinct groups; for example, categorical predictors include gender, material type, and payment method. 'Continuous variables' are numeric variables with an infinite number of values between any two points, such as the length of a part or the date and time a payment is received.

#### Categorical variables

We have 7 categorical variables in the data:

+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Variable Name | Description                                                                                                                                                                                                                                                                                                                                                                             |
+===============+=========================================================================================================================================================================================================================================================================================================================================================================================+
| sex           | sex (male or female)                                                                                                                                                                                                                                                                                                                                                                    |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| cp            | The 'cp' variable categorizes chest pain type in the dataset, with values representing whether it's typical angina, atypical angina, non-anginal pain, or asymptomatic. Angina refers to chest pain or discomfort caused by reduced blood flow to the heart muscles, and understanding its type provides valuable insights into patients symptomatic profiles for clinical assessments. |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| fbs           | The 'fbs' variable signals whether fasting blood sugar exceeds 120 mg/dl (true) or not (false), providing a rapid assessment of hyperglycemia, which is an elevated level of glucose in the blood after a period of fasting.                                                                                                                                                            |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| restecg       | The 'restecg' variable characterizes resting electrocardiographic results with values indicating:                                                                                                                                                                                                                                                                                       |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Normal:** No significant abnormalities.                                                                                                                                                                                                                                                                                                                                           |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **ST-T Wave Abnormality:** Presence of T wave inversions and/or ST elevation or depression of \> 0.05 mV, suggesting potential cardiac issues.                                                                                                                                                                                                                                      |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Probable or Definite Left Ventricular Hypertrophy by Estes' criteria:** Indicative of an enlarged left ventricle, determined by specific criteria.                                                                                                                                                                                                                                |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | This variable efficiently communicates essential cardiac indicators, facilitating the identification of electrocardiographic abnormalities in the dataset. Resting electrocardiography is a non-invasive test measuring the electrical activity of the heart at rest, providing valuable insights into cardiac health.                                                                  |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| exang         | The 'exang' variable denotes the presence of exercise-induced angina:                                                                                                                                                                                                                                                                                                                   |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Yes:** Indicates the occurrence of exercise-induced angina.                                                                                                                                                                                                                                                                                                                       |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **No:** Signifies the absence of exercise-induced angina.                                                                                                                                                                                                                                                                                                                           |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | This variable provides a straightforward representation of the presence or absence of angina triggered by physical activity within the dataset.                                                                                                                                                                                                                                         |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| slope         | The 'slope' variable characterizes the slope of the peak exercise ST segment, with values indicating:                                                                                                                                                                                                                                                                                   |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Upsloping:** A rising ST segment during peak exercise, often considered a normal response.                                                                                                                                                                                                                                                                                        |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Flat:** The ST segment maintains a consistent level during peak exercise, which may suggest a borderline or intermediate response.                                                                                                                                                                                                                                                |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Downsloping:** A descending ST segment during peak exercise, potentially indicating an abnormal cardiac response, which could be associated with ischemia.                                                                                                                                                                                                                        |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | This variable succinctly communicates the pattern of the ST segment during peak exercise, offering insights into the cardiac response under stress within the dataset.                                                                                                                                                                                                                  |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| thal          | The 'thal' variable shows different results from a special heart test called thallium stress testing:                                                                                                                                                                                                                                                                                   |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Normal:** Nothing unusual found during the test, where they use thallium-201 to check how blood flows to the heart.                                                                                                                                                                                                                                                               |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Fixed Defect:** Points to a lasting issue seen in the thallium scan.                                                                                                                                                                                                                                                                                                              |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Reversible Defect:** Suggests a problem seen in the thallium scan that might get better, often linked to not enough blood flow.                                                                                                                                                                                                                                                   |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | This variable helps us understand various outcomes from thallium stress testing, showing potential heart issues in the dataset. Thallium stress testing is a test that uses a bit of thallium-201, a special substance, to see how well blood is moving through the heart.                                                                                                              |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| diagnosed     | The 'num' variable serves as the basis for diagnosing heart disease (angiographic disease status) and is interpreted as follows:                                                                                                                                                                                                                                                        |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **Yes:** Denoting the presence of heart disease.                                                                                                                                                                                                                                                                                                                                    |
|               |                                                                                                                                                                                                                                                                                                                                                                                         |
|               | -   **No:** Signifying the absence of heart disease.                                                                                                                                                                                                                                                                                                                                    |
+---------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+

### Numerical variables

### Continuous variables

+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Variable Name | Description                                                                                                                                                                                                                                                                                                                                                                                                |
+===============+============================================================================================================================================================================================================================================================================================================================================================================================================+
| age           | age in years                                                                                                                                                                                                                                                                                                                                                                                               |
+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| trestbps      | resting blood pressure (in mm Hg on admission to the hospital) measured when patients were admitted to the hospital, giving us a baseline indicator of their cardiovascular health.                                                                                                                                                                                                                        |
+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| chol          | The 'chol' variable represents serum cholesterol levels measured in milligrams per deciliter (mg/dL). This value shows how much cholesterol is in the blood, which is a key factor in assessing cardiovascular health within the dataset. Cholesterol in the blood is a measure of the fats in the body and can impact heart health; higher levels may indicate a potential risk for heart-related issues. |
+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| thalach       | The 'thalach' variable represents the highest heart rate reached during a cardiovascular stress test. This number shows how well the heart can work and is an important measure for evaluating heart health in the dataset.                                                                                                                                                                                |
+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| oldpeak       | The 'oldpeak' number shows how much the ST segment in the heart's electrical signal drops during exercise compared to when at rest. This drop is called ST depression and helps us understand how the heart responds to physical activity, indicating potential issues with blood flow in the dataset.                                                                                                     |
+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| ca            | The 'ca' number shows how many major blood vessels, ranging from 0 to 3, have been colored during a special X-ray test. This helps us understand if there are any blockages or issues in the big blood vessels, giving important information about heart health in the dataset.                                                                                                                            |
+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| num           | The 'num' tells us if there might be a problem with the heart:                                                                                                                                                                                                                                                                                                                                             |
|               |                                                                                                                                                                                                                                                                                                                                                                                                            |
|               | -   **Value 0:** Means the heart's arteries are not much narrowed.                                                                                                                                                                                                                                                                                                                                         |
|               |                                                                                                                                                                                                                                                                                                                                                                                                            |
|               | -   **Values 1, 2, and 3:** All mean the heart's arteries are more narrowed, suggesting a potential problem.                                                                                                                                                                                                                                                                                               |
|               |                                                                                                                                                                                                                                                                                                                                                                                                            |
|               | In simple words, 'num' helps us figure out if there's a significant narrowing in the heart arteries, which could be a sign of a heart issue. If the number is 0, it's less severe, and if it's 1, 2, or 3, it suggests a more serious narrowing.                                                                                                                                                           |
+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
