==> R/05_analysis_1.qmd <==
---
title: "Analysis 1"
format:
  html: {}
---

# Analysis of data

### Libraries

```{r}
library(ggplot2)
library(tidyverse)
library(patchwork)
library(ggpubr)
library(broom)
source("99_proj_func.R")
```

### Loading data

```{r}
#| message: false
data <- read_tsv("../data/03_dat_aug.tsv")
```

## Data Overview

### **Exploring Age Distribution by Diagnosis**

First, we generate a violin plot graphing the diagnosis of being healthy or sick against age. Which provides a quick and insightful snapshot for initial dataset exploration.

```{r}
#Violin plot
violin_p1 <- ggplot(data = data,
                 aes(x = age,
                     y = diagnose,
                     color = diagnose,
                     fill = diagnose)) +
  geom_violin(alpha = 0.4) +
  scale_fill_manual(values = c("darkgreen", "red")) +
  scale_color_manual(values = c("darkgreen", "red")) +
  theme_minimal()

violin_p1
ggsave(file = "../results/plot_1.png", violin_p1, width = 8, height = 6, units = "in")
```

The x-axis represents the age of an individual, ranging from 30 to 70. The y-axis shows the relative amount of diagnosis of heart disease, labeled "sick" or "healthy". Each diagnosis is visually differentiated by two colors. The "green" color represents a positive health condition ("healthy"), while the "red" color marks a negative health condition ("sick"). On one hand, the red curve is broader, peaking around the age of 50. On the other hand, the green curve is narrower, also peaking around the age of 50 but with a lower amplitude compared to the red curve. In summary, the plot shows us that the age-attribute is not evenly distributed between healthy and sick individuals. Most diagnosed individuals were between the age of 55 and 65. The highest proportion of healthy individuals can be noticed between the approximate ages of 40 and 55, though relatively evenly distributed for the small sample.

### **Exploring Numerical Variables by Diagnosis**

Before searching for more detailed correlations between variables, we explore the numerical data through density plots. The goal is to create an overview of their distributions and how well they might separate healthy from sick patients.

```{r}
#Plotting the distribution of numerical variables, divided by diagnosis

# Density-plot of age attribute
density_p1 <- dens(data, "age")

# Density-plot of trestbps attribute
density_p2 <- dens(data, "trestbps")

# Density-plot of chol attribute
density_p3 <- dens(data, "chol")

# Density-plot of thalach attribute
density_p4 <- dens(data, "thalach")

# Density-plot of ca attribute
density_p5 <- dens(data, "ca")

# Density-plot of oldpeak attribute
density_p6 <- dens(data, "oldpeak")

# All density plots
density_grid <- ggarrange(density_p1, density_p2, density_p3, density_p4, density_p5, density_p6, ncol=2, nrow=3, common.legend = TRUE, legend = "bottom")

density_grid
ggsave(file = "../results/plot_2.png", density_grid, width = 8, height = 6, units = "in")

```

The result consists of six separate density plots arranged in a grid. Each plot focuses on the distribution of individual variables. As in the violin plot, the difference in the diagnosis is represented by the colors 'green' or 'red', indicating whether the individual is healthy or sick.

In the "age" plot, the green curve has a peak around age 55 while the red curve peaks around age 60, which is also what we saw from the violin plot.

In the "trestbps" plot, the curves overlap substantially. Both have a peak around 130 but the red curve is slightly shifted right indicating higher blood pressure in sick individuals.

In the "chol" plot, the curves seem quite similar again. The green curve peaks around 240, while the red curve peaks at about 260, indicating higher cholesterol levels in sick persons.

In the "thalach" plot, the curves show a greater difference. The green curve peaks near 160, while the red one peaks around 140, showing that ill persons tend to have a lower maximum heart rate.

The "ca" plot shows that people sick with heart disease often have one or more colored blood vessels from performing the X-ray test. The difference between the distribution of healthy and sick people is especially prevalent for this plot compared to the others, indicating that it could be an insightful attribute for further analysis.

The last plot is of the "oldpeak" variable, which also shows great promise in separating sick from healthy individuals. It shows that a large drop in electrical signal of the ST segment could be a strong indicator of heart disease.

To further investigate the relationship between the diagnosis of a patient and the numerical data, each attribute was fitted to a linear model.

```{r}
# Elongating the dataframe
model_data <- data |>
  select(diagnose, age, trestbps, chol, thalach, ca, oldpeak) |>
  mutate(diagnose = case_when(diagnose == "healthy" ~ 0,
                              diagnose == "sick" ~ 1)) |>
  pivot_longer(cols = c(age, trestbps, chol, thalach, ca, oldpeak),
               names_to = "numerical_variables",
               values_to = "numerical_value")

# Fitting the data to a linear model
fitted_data <- model_data |>
  group_by(numerical_variables) |>
  nest() |>
  mutate(linear_model = map(.x = data,
                            .f = ~lm(numerical_value ~ diagnose,
                                     data = .x))) |> 
  mutate(linear_model_tidy = map(.x = linear_model,
                                 .f = ~tidy(x = .x,
                                            conf.int = TRUE,
                                            conf.level = 0.95))) |>
  unnest(linear_model_tidy)

# Getting the p-value of each model
fitted_data |>
  filter(term == "diagnose") |>
  select(numerical_variables, p.value) |>
  arrange(p.value)

```

Looking at the order of the p-values acquired from the linear regression, we can draw a similar conclusion to what was drawn by looking at the density plots. It shows a linear correlation between the diagnosis of a patient and the 'ca', 'oldpeak', and 'thalach' variables.

### **Exploring Categorical Variables by Diagnosis**

In the following, we create a series of bar plots examining the distribution of various categorical variables in relation to the diagnosis of a patient.

```{r}
# Plotting the disctribution of categorical variables, divided by diagnosis

# Bar-plot of cp attribute
bar_p1 <- bar(data, "cp")

# Bar-plot of thal attribute
bar_p2 <- bar(data, "thal")

# Bar-plot of fbs attribute
bar_p3 <- bar(data, "fbs")

# Bar-plot of exang attribute
bar_p4 <- bar(data, "exang")

# Bar-plot of restecg attribute
bar_p5 <- bar(data, "restecg")

# Bar-plot of slope attribute
bar_p6 <- bar(data, "slope")

# All bar plots
bar_grid <- ggarrange(bar_p1, bar_p2, bar_p3, bar_p4, bar_p5, bar_p6, ncol = 2, nrow = 3, common.legend = TRUE, legend = "bottom")
bar_grid
ggsave(file = "../results/plot_3.png", bar_grid, width = 8, height = 6, units = "in")

```

The plots compares the health status (healthy or sick) of individuals based on various categorical variables, i.e. chest pain type ('cp'), fasting blood sugar ('fbs'), resting electrocardiographic results ('restagc'), exercise induced angina ('exang'), thallium heart scan ('thal'), and the slope peak exercise ST segment ('slope').

The "cp" plot shows that people with asymptomatic chest pain often are sick, while individuals with atypical angina, non-anignal pain, and typical angina, are mostly healthy.

In the "thal" plot, those with fixed or reversible defects are mostly sick, while those with normal results are healthy.

The "fbs" plot shows that fasting blood sugar doesn't make a huge difference if an individual is diagnosed as either healthy or sick, as the bars with lower than 120mg/ml and greater than 120mg/ml have similar proportions.

In the "exang" plot, it's shown that the individuals have an occurrence of exercise-induced angina are primarly sick. Those with an absence of exercise-induced angina are largely healthy.

For "restecg", individuals with ST-T wave abnormality are most likely sick. Others in this category lean towards being diagnosed as healthy. People having left ventricular hyperthrophy have a equal proportion of being either healthy or sick.

The "slope" plot shows that an upsloping ST segment could be an indicator of health, while a flat or downturned slope could indictae heart disease.

From all six plots, it seems like types of chest pain could be the best attribute to indicate heart disease in a patient, though the thallium heart scan and exercise induced angina also separate the two groups well.

# Discussion

From the visualization and linear models, we've formed a general understanding of our data, and the relationship between the diagnosis of a patient, and all other attributes. The numerical variables which showed the strongest correlation to diagnosis were 'ca', 'oldpeak', and 'thalach', while the best categorical variables were 'cp', 'thal', and 'exang'.
